+++ wantedcqrs/src/main/java/com/pawn/wantedcqrs/optionGroup/dto/ProductOptionGroupDto.java
@@ -1,6 +1,6 @@
-package com.pawn.wantedcqrs.productOptionGroup.dto;
+package com.pawn.wantedcqrs.optionGroup.dto;
 
-import com.pawn.wantedcqrs.productOptionGroup.entity.ProductOptionGroup;
+import com.pawn.wantedcqrs.optionGroup.entity.ProductOptionGroup;
 import lombok.Data;
 
 import java.util.ArrayList;
