+++ wantedcqrs/src/main/java/com/pawn/wantedcqrs/product/dto/CreateProductRequest.java
@@ -2,13 +2,12 @@ package com.pawn.wantedcqrs.product.dto;
 
 import com.pawn.wantedcqrs.product.entity.Currency;
 import com.pawn.wantedcqrs.product.entity.ProductStatus;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductOptionDto;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductOptionGroupDto;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductOptionDto;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductOptionGroupDto;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.math.BigDecimal;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
