+++ wantedcqrs/src/main/java/com/pawn/wantedcqrs/product/service/ProductAssembler.java
@@ -4,8 +4,8 @@ import com.pawn.wantedcqrs.brand.dto.BrandDto;
 import com.pawn.wantedcqrs.category.dto.CategoryDto;
 import com.pawn.wantedcqrs.product.dto.*;
 import com.pawn.wantedcqrs.product.repository.dto.ProductSummaryProjection;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductOptionGroupDto;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductStockProjection;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductOptionGroupDto;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductStockProjection;
 import com.pawn.wantedcqrs.review.dto.ReviewStatsProjection;
 import com.pawn.wantedcqrs.review.repository.dto.ReviewDistributionProjection;
 import com.pawn.wantedcqrs.seller.dto.SellerDto;
