+++ wantedcqrs/src/main/java/com/pawn/wantedcqrs/product/service/ProductFacade.java
@@ -7,9 +7,9 @@ import com.pawn.wantedcqrs.category.service.CategoryService;
 import com.pawn.wantedcqrs.common.exception.e4xx.ResourceNotFoundException;
 import com.pawn.wantedcqrs.product.dto.*;
 import com.pawn.wantedcqrs.product.repository.dto.ProductSummaryProjection;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductOptionGroupDto;
-import com.pawn.wantedcqrs.productOptionGroup.dto.ProductStockProjection;
-import com.pawn.wantedcqrs.productOptionGroup.service.ProductOptionGroupService;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductOptionGroupDto;
+import com.pawn.wantedcqrs.optionGroup.dto.ProductStockProjection;
+import com.pawn.wantedcqrs.optionGroup.service.ProductOptionGroupService;
 import com.pawn.wantedcqrs.review.dto.ReviewStatsProjection;
 import com.pawn.wantedcqrs.review.repository.dto.ReviewDistributionProjection;
 import com.pawn.wantedcqrs.review.service.ReviewService;
